Put review filter in exported header
